pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-platform-sdk"
version = "0.1.1"
description = "QMT trading SDK with WebSocket (Centrifugo) integration"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "centrifuge-python>=0.4.3",
    "aiohttp>=3.9.0",
    "pandas>=1.5.3",
]

[tool.setuptools.packages.find]
where = ["src"]
